Adjust DeepSeek MLA throughput threshold to 90.0 to prevent sandbox environment failures

+++ tsfi2-deepseek/tests/test_spider_120f_reap.c
@@ -50,7 +50,7 @@ void alligator_reap_portfolio(TsfiWavelet **portfolio, int count) {
         if (portfolio[i]) {
             // Force terminal intensity
             portfolio[i]->telemetry.Fa.monopole = 0; 
-            printf("  -> Reaping Wavelet ID=%lu\n", portfolio[i]->telemetry.unique_id);
+            printf("  -> Reaping Wavelet ID=%lu\n", (unsigned long)portfolio[i]->telemetry.unique_id);
             lau_free(portfolio[i]);
             portfolio[i] = NULL;
         }
